Divided the CgiOut script into multiple scripts
